+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added parameter `is_raven` to load function to specify whether a custom Protobuf model is a Raven model or not
+- Added tests
+
 ## [0.2.0] - 2025-10-27
 
 ### Changed